added some changes for delete

+++ BlizuTebe/Services/LocalCommunityService.cs
@@ -1,6 +1,7 @@
 ﻿using AutoMapper;
 using BlizuTebe.Dtos;
 using BlizuTebe.Models;
+using BlizuTebe.Repositories;
 using BlizuTebe.Repositories.Interfaces;
 using BlizuTebe.Services.Interfaces;
 using FluentResults;
@@ -15,12 +16,20 @@ namespace BlizuTebe.Services
         private readonly ILocalCommunityRepository _repository;
         private readonly GeoJsonReader _geoJsonReader;
         private readonly GeoJsonWriter _geoJsonWriter;
+        private readonly IUserRepository _userRepository;
+        private readonly IAnnouncementRepository _announcementRepository;
+        private readonly ICommunityRequestRepository _communityRequestRepository;
+        private readonly IDiscussionRepository _discussionRepository;
 
-        public LocalCommunityService(ILocalCommunityRepository repository)
+        public LocalCommunityService(ILocalCommunityRepository repository, ICommunityRequestRepository communityRequestRepository, IDiscussionRepository discussionRepository, IUserRepository userRepository, IAnnouncementRepository announcementRepository)
         {
             _repository = repository;
             _geoJsonReader = new GeoJsonReader();
             _geoJsonWriter = new GeoJsonWriter();
+            _userRepository = userRepository;
+            _announcementRepository = announcementRepository;
+            _communityRequestRepository = communityRequestRepository;
+            _discussionRepository = discussionRepository;
         }
 
         public Result<LocalCommunityDto> Create(LocalCommunityDto dto)
@@ -52,6 +61,37 @@ namespace BlizuTebe.Services
             return Result.Ok(MapToDto(community)); // Ručna konverzija
         }
 
+        public Result<LocalCommunityDto> Update(long id, LocalCommunityDto dto)
+        {
+            var community = _repository.GetById(id);
+            if (community == null)
+            {
+                return Result.Fail<LocalCommunityDto>("Mesna zajednica nije pronađena sa ID: " + id);
+            }
+
+            community.Name = dto.Name;
+            community.City = dto.City;
+            community.PresidentId = dto.PresidentId;
+            community.PhoneNumber = dto.PhoneNumber;
+            community.Facebook = dto.Facebook;
+
+            if (!string.IsNullOrEmpty(dto.Boundary))
+            {
+                community.Boundary = _geoJsonReader.Read<Geometry>(dto.Boundary);
+                community.Boundary.SRID = 4326;
+            }
+
+            if (dto.CenterPoint != null && dto.CenterPoint.Length == 2)
+            {
+                var factory = new GeometryFactory(new PrecisionModel(), 4326);
+                community.CenterPoint = factory.CreatePoint(new Coordinate(dto.CenterPoint[0], dto.CenterPoint[1]));
+            }
+
+            _repository.Update(community);
+
+            return Result.Ok(MapToDto(community));
+        }
+
         public Result<LocalCommunityDto> GetById(long id)
         {
             var community = _repository.GetById(id);
@@ -76,6 +116,14 @@ namespace BlizuTebe.Services
             if (community == null)
             {
                 return Result.Fail<LocalCommunityDto>("Mesna zajednica nije pronađena sa ID: " + id);
+            }
+
+            if (_userRepository.ExistsByCommunityId(id)
+            || _discussionRepository.ExistsByCommunityId(id)
+            || _communityRequestRepository.ExistsByCommunityId(id)
+            || _announcementRepository.ExistsByCommunityId(id))
+            {
+                return Result.Fail<LocalCommunityDto>("Brisanje nije dozvoljeno jer postoje povezani entiteti.");
             }
 
             _repository.Delete(community);
